requirements/testing.txt
# Testing dependencies
# --------------------

pytest-cov>=4
packaging>=23
pytest-qgis>=2.1.0
